requirements.txt
flask-socketio==5.4.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
gunicorn==23.0.0
eventlet==0.37.0
orjson==3.10.12
bleach==6.2.0